Move the info page into an Wouter page
