+++ package.json
@@ -24,7 +24,9 @@
     "plasmo": "^0.84.2",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "webextension-polyfill": "^0.10.0"
+    "react-router-dom": "^6.22.1",
+    "webextension-polyfill": "^0.10.0",
+    "wouter": "^2.12.1"
   },
   "devDependencies": {
     "@ianvs/prettier-plugin-sort-imports": "4.1.0",
