+++ popup/util/page.ts
@@ -1,0 +1,5 @@
+export enum PopupPath {
+  root = "/",
+  setting = "/setting",
+  info = "/info"
+}
